% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton_status.R
\name{merton_status}
\alias{merton_status}
\title{Check Merton Server Status}
\usage{
merton_status()
}
\value{
A list with version information if the server is reachable. An error is thrown otherwise.
}
\description{
Sends a request to the \verb{/api/version} endpoint of the Merton (Ollama) server to check if the server is reachable and authorised.
}
\examples{
merton_status()
}
